Listen and forget
This is the 'listen and forget' version. Since no data is stored in any array, there is not a playlist or a 'previous track' button. Contains the native HTML5 audio player of browser. As soon as current track ends and/or user jumps to next track, script makes an AJAX request for the next random one related to the current track.